refactor, split out custom event templates service
