+++ src/main/java/io/cryostat/events/EventTemplates.java
@@ -15,62 +15,32 @@
  */
 package io.cryostat.events;
 
-import java.io.IOException;
 import java.net.URI;
-import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
-import org.openjdk.jmc.common.unit.IConstrainedMap;
-import org.openjdk.jmc.flightrecorder.configuration.events.EventOptionID;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.JFCGrammar;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.XMLAttributeInstance;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.XMLModel;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.XMLTagInstance;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.XMLValidationResult;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.model.EventConfiguration;
-
 import io.cryostat.ConfigProperties;
-import io.cryostat.core.FlightRecorderException;
-import io.cryostat.core.templates.MutableTemplateService.InvalidEventTemplateException;
-import io.cryostat.core.templates.MutableTemplateService.InvalidXmlException;
 import io.cryostat.core.templates.Template;
-import io.cryostat.core.templates.TemplateService;
 import io.cryostat.core.templates.TemplateType;
 import io.cryostat.targets.Target;
 import io.cryostat.targets.TargetConnectionManager;
-import io.cryostat.util.HttpStatusCodeIdentifier;
 
-import io.quarkus.runtime.StartupEvent;
-import io.smallrye.common.annotation.Blocking;
 import io.smallrye.mutiny.Uni;
 import io.vertx.core.Vertx;
 import jakarta.annotation.security.RolesAllowed;
-import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.core.Response;
-import org.apache.hc.core5.http.ContentType;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.reactive.RestForm;
 import org.jboss.resteasy.reactive.RestPath;
 import org.jboss.resteasy.reactive.RestResponse;
 import org.jboss.resteasy.reactive.multipart.FileUpload;
-import org.jsoup.nodes.Document;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
-import software.amazon.awssdk.services.s3.model.HeadBucketRequest;
-import software.amazon.awssdk.services.s3.model.ListObjectsV2Request;
-import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 @Path("")
 public class EventTemplates {
@@ -86,37 +56,11 @@ public class EventTemplates {
 
     @Inject Vertx vertx;
     @Inject TargetConnectionManager connectionManager;
-    @Inject S3Client storage;
+    @Inject S3TemplateService customTemplateService;
     @Inject Logger logger;
 
     @ConfigProperty(name = ConfigProperties.AWS_BUCKET_NAME_EVENT_TEMPLATES)
     String eventTemplatesBucket;
-
-    void onStart(@Observes StartupEvent evt) {
-        // FIXME refactor this to a reusable utility method since this is done for custom event
-        // templates, archived recordings, and archived reports
-        boolean exists = false;
-        try {
-            exists =
-                    HttpStatusCodeIdentifier.isSuccessCode(
-                            storage.headBucket(
-                                            HeadBucketRequest.builder()
-                                                    .bucket(eventTemplatesBucket)
-                                                    .build())
-                                    .sdkHttpResponse()
-                                    .statusCode());
-        } catch (Exception e) {
-            logger.info(e);
-        }
-        if (!exists) {
-            try {
-                storage.createBucket(
-                        CreateBucketRequest.builder().bucket(eventTemplatesBucket).build());
-            } catch (Exception e) {
-                logger.error(e);
-            }
-        }
-    }
 
     @GET
     @Path("/api/v1/targets/{connectUrl}/templates")
@@ -140,7 +84,7 @@ public class EventTemplates {
                 .onComplete(
                         ar -> {
                             try {
-                                addTemplate(ar.result().toString());
+                                customTemplateService.addTemplate(ar.result().toString());
                                 cf.complete(null);
                             } catch (Exception e) {
                                 logger.error(e);
@@ -202,99 +146,5 @@ public class EventTemplates {
                                 .getXml(templateName, templateType)
                                 .orElseThrow(NotFoundException::new)
                                 .toString());
-    }
-
-    static class S3TemplateService implements TemplateService {
-        S3Client s3;
-
-        @Override
-        public Optional<IConstrainedMap<EventOptionID>> getEvents(
-                String templateName, TemplateType templateType) throws FlightRecorderException {
-            return Optional.empty();
-        }
-
-        @Override
-        public List<Template> getTemplates() throws FlightRecorderException {
-            var builder = ListObjectsV2Request.builder().bucket(eventTemplatesBucket);
-            var objects = s3.listObjectsV2(builder.build());
-            var templates = convertObjects(objects);
-            return templates;
-        }
-
-        private List<Template> convertObjects(ListObjectsV2Response objects) {
-            return List.of();
-        }
-
-        @Override
-        public Optional<Document> getXml(String templateName, TemplateType templateType)
-                throws FlightRecorderException {
-            return Optional.empty();
-        }
-    }
-
-    @Blocking
-    public Template addTemplate(String templateText)
-            throws InvalidXmlException, InvalidEventTemplateException, IOException {
-        try {
-            XMLModel model = EventConfiguration.createModel(templateText);
-            model.checkErrors();
-
-            for (XMLValidationResult result : model.getResults()) {
-                if (result.isError()) {
-                    // throw new InvalidEventTemplateException(result.getText());
-                    throw new IllegalArgumentException(result.getText());
-                }
-            }
-
-            XMLTagInstance configuration = model.getRoot();
-            XMLAttributeInstance labelAttr = null;
-            for (XMLAttributeInstance attr : configuration.getAttributeInstances()) {
-                if (attr.getAttribute().getName().equals("label")) {
-                    labelAttr = attr;
-                    break;
-                }
-            }
-
-            if (labelAttr == null) {
-                // throw new InvalidEventTemplateException(
-                //         "Template has no configuration label attribute");
-                throw new IllegalArgumentException("Template has no configuration label attribute");
-            }
-
-            String templateName = labelAttr.getExplicitValue();
-            templateName = templateName.replaceAll("[\\W]+", "_");
-
-            XMLTagInstance root = model.getRoot();
-            root.setValue(JFCGrammar.ATTRIBUTE_LABEL_MANDATORY, templateName);
-
-            String key = templateName;
-            storage.putObject(
-                    PutObjectRequest.builder()
-                            .bucket(eventTemplatesBucket)
-                            .key(key)
-                            .contentType(ContentType.APPLICATION_XML.getMimeType())
-                            .build(),
-                    RequestBody.fromString(model.toString()));
-
-            return new Template(
-                    templateName,
-                    getAttributeValue(root, "description"),
-                    getAttributeValue(root, "provider"),
-                    TemplateType.CUSTOM);
-        } catch (IOException ioe) {
-            // throw new InvalidXmlException("Unable to parse XML stream", ioe);
-            throw new IllegalArgumentException("Unable to parse XML stream", ioe);
-        } catch (ParseException | IllegalArgumentException e) {
-            // throw new InvalidEventTemplateException("Invalid XML", e);
-            throw new IllegalArgumentException("Invalid XML", e);
-        }
-    }
-
-    protected String getAttributeValue(XMLTagInstance node, String valueKey) {
-        return node.getAttributeInstances().stream()
-                .filter(i -> Objects.equals(valueKey, i.getAttribute().getName()))
-                .map(i -> i.getValue())
-                .findFirst()
-                .get();
     }
 }
